backend/requirements.txt
Django
psycopg2
djangorestframework
django-cors-headers
djangorestframework-simplejwt
djoser